Moved logger classes to connect.logger package.
