+++ connect/logger/Logger.hx
@@ -1,4 +1,4 @@
-package connect;
+package connect.logger;
 
 
 /**
