+++ connect/logger/MarkdownLoggerFormatter.hx
@@ -1,4 +1,4 @@
-package connect;
+package connect.logger;
 
 
 @:dox(hide)
