+++ examples/Example.hx
@@ -1,8 +1,8 @@
 import connect.Env;
 import connect.Flow;
-import connect.Logger;
 import connect.Processor;
 import connect.api.Query;
+import connect.logger.Logger;
 
 
 class Example {
@@ -11,7 +11,7 @@ class Example {
         var initialTime = Date.now().getTime();
 
         Env.loadConfig('examples/config.json');
-        Env.initLogger('examples/log.md', Logger.LEVEL_INFO, null, null);
+        Env.initLogger('examples/log', Logger.LEVEL_INFO, null, null);
 
         // Define main flow
         var flow = new Flow(null)
