+++ examples/example.js
@@ -1,11 +1,11 @@
 const connect = require("../_packages/connect.js/connect");
 const Env = connect.Env;
 const Flow = connect.Flow;
-const Logger = connect.Logger;
 const Processor = connect.Processor;
 const Query = connect.api.Query;
+const Logger = connect.logger.Logger;
 
-//Env.initLogger("log.md", Logger.LEVEL_ERROR, null, null);
+Env.initLogger("log", Logger.LEVEL_INFO, null, null);
 
 // Define main flow
 const flow = new Flow(null)
